test: separate snapshot testing
